Replace time-based connection recycling with request counter
The time-based force_close() after N seconds didn't effectively limit
buffer growth because busy connections never idle.  Instead, count
requests per connection and force_close() after max_conn_lifetime_secs
requests (reusing the config field name for compatibility).

With max_conn_lifetime_secs=100, each connection handles ~100 requests
(~7 seconds at 46K req/min across 50 connections) before recycling.
This creates ~420 connection recycles/min (vs 46K/min with keepalive
off), keeping TIME_WAIT under 500 vs 46K, while still preventing
BytesMut buffer accumulation.

Uses AtomicU32 per-connection counter in on_connect extension data.

+++ src/config.rs
@@ -79,11 +79,12 @@ pub struct DroneConfig {
     #[serde(default)]
     pub debug_endpoints_enabled: bool,
 
-    /// Maximum connection lifetime in seconds (0 = unlimited).  After this many seconds,
-    /// responses include "Connection: close" to force the client to reconnect.  This works
-    /// around actix-http's BytesMut buffers growing monotonically on long-lived connections.
+    /// Maximum requests per connection before forcing close (0 = unlimited).
+    /// After this many requests, drone sends force_close() to recycle the connection,
+    /// freeing actix-http's BytesMut buffers that grow monotonically.
+    /// Recommended: 100 (~7 seconds at typical load).
     #[serde(default)]
-    pub max_conn_lifetime_secs: u64,
+    pub max_conn_lifetime_secs: u32,
 }
